Add a test that checks when the allocator starts spilling.
This test pinpointed a subtelty: Even though RDI is reserved for trace
inputs, upon its last use, we still mark it free, thus allowing another
local to use the register.

This behaviour is fine (and actually good for performance), but I wanted
to add a comment to make it super-explicit, as it may be unexpected.