Add TraceCompiler::new() and inline TraceCompiler::init().
We have multiple places in the code doing the (somewhat lengthy)
initialisation of a trace compiler. Adding a new() therefore make
sense and removes some duplication.

Then we have an init() method, which is called in just one place.
There's no need for it to be separate from new(), so this change inlines
it.