Test "direct-to-stack" register allocation cases.
There are a handful of cases in the register allocator where a local
will be allocated on the stack even if registers are available. Some of
the cases are intentional, others are this way just to get us going.

Regardless, we should test these invariants so that we notice if they no
longer hold.