Handle threads that don't finish tracing gracefully.
Previously, if we started tracing but didn't close the loop before the thread
finished, the Location set to PHASE_TRACING would be stuck in that phase
forever.

This commit changes that. When a thread comes across a PHASE_TRACING Location,
it now checks to see if the associated thread is still alive. If it is not, it
changes the Location to PHASE_WONT_TRACE which means we'll never try tracing the
Location again. This latter phase is a temporary hack: in the longer run, we
need to do "retry N times before marking as WONT_TRACE" (as RPython does). Doing
that is for a future commit.

The cunning hack in this commit is how we detect if the tracing thread is alive:
we just change the thread ID we use to be a pointer to an Arc (where the
contents of the Arc aren't important). If that Arc's strong count is above 1,
the other thread is alive, otherwise it's finished.