Only push to the return stack if the call is to be inlined.
This change fixes a bug where we'd update the return stack for external
calls (for example, things marked `#[do_not_trace]`). This is wrong
because we will never see a corresponding Return terminator for the
callee, and then some poor unrelated (and unsuspecting) Return ends up
popping a garbage value off of the return stack.

Fixes #188.