Remove an old ignored test that no longer makes sense.
This test covers a behaviour no longer present in the allocator: we no
longer evict locals from registers. This was removed during the SIR
redesign, as it turned out to be problematic (it was very easy to
accidentally use stale Locations, because a variable's allocation could
have changed behind your back).