Remove more problematic ignored tests.
These tests attempt to guess the stack allocation of a trace based on
the number or rust-level local variables used in the traced function.

In reality it's very difficult to know how much stack will be used due
to the use of intermediate stores in the trace. Even if we were to
hard-code the stack height we see today, the test is likely to be
extremely fragile and arguably it doesn't test anything useful any more.