Add PHASE_TRACING_LOCK to prevent accessing freed memory.
I have never observed a crash here, but it's clear that one could happen,
because the first thing we did in PHASE_TRACING previously was to reconstruct a
box, which another thread could free while we thought it was still alive.

This commit introduces a new PHASE_TRACING_LOCK which is a lock that only
PHASE_TRACING uses. It means that we can't read possibly-freed memory. It does
mean that we introduce two spinlock phases, both of which will have terrible
performance in the case of many threads trying to trace the same location.
However, we can fix the performance later, perhaps using a futex or something
similar.

It might be tempting to think we can reuse PHASE_LOCKED here, as I spent several
hours thinking. However, PHASE_LOCKED allows a thread to continue executing when
it observes it, whereas PHASE_TRACING_LOCK stops all threads that observe it
until it changes. The reason for this is that if the current trace is tracing
the current location, we can't carry on execution, or the resultant trace will
contain more than 1 iteration, with consequent hilarity.