Reinstate ignored array slice test.
This test was checking that a) MIR `Len` rvalues are correctly lowered
to SIR, and that b) the JIT correctly accesses the length in the second
part of the slice's fat pointer (as part of an array indexing operation).

I've managed to fix the test and simplify it in doing so.

I also found a bug in `copy_memory`, where loading the first operand to
memmove(3) overwrote the place we'd load the second operand from.