Remove Location::NotLive.
TraceCompiler::variable_location_map keeps track of the register
allocation of local variables.

We currently use a Location::NotLive variant to express that an variable
is not live and thus has no allocation. But we can instead just not
store that key in the variable_location_map.

This has the upshot of not having to continually check for
Location::NotLive (which, if we have done things correctly, can't arise
anyway) when matching over a Location.