Use MIR liveness analysis instead of our own.
We previously ran our own liveness analysis to be able to remove
variables and statements that weren't relevant for the TIR trace.
However, for blackholing we now need to keep the variables and
statements around as they may be needed in the SIR. This means we can go
back to using MIRs liveness analysis to compute live variables for the
guards.