Implement blackholing.
If during the execution of a compiled trace a guard fails, execution
needs to resume in the interpreter. Unfortunately, guards may fail in
the middle of a trace, and we can't just pick up normal interpretation
from there, since the interpreter won't know about any values computed
up to this point. Due to side effects we also can't just "undo" the
execution of compiled trace and re-run that code in the interpreter.
Instead, we need to run the branches that weren't compiled to get back
to a control point from which normal interpretation can continue. This
is sometimes, in lack of a better term, called blackholing.

In our case, blackholing is done using a special interpreter, the
SIRInterpreter, which runs on SIR blocks. When a guard fails, we check
which block in the SIR execution needs to continue. We then initialise
the SIRInterpreter at that point, copying over all live values computed
during execution of the compiled trace, and then interpret SIR blocks
until we get back to the control point (which happens when we return
from the `interp_step`-annotated function).

Since the SIRInterpreter interprets SIR statements, which assume
variables live in memory and are often referenced via offsets, the
interpreter reads from and writes to allocated memory blocks. In order
to initialise these blocks with currently live variables, they need to
be allocated within the compiled trace upon guard failure. Each guard
stores information about which variables are live up to that point.
Guards also need to store information about function calls (even though
in the compiled trace those are inlined), so that we can recreate the
needed stack frames in the SIRInterpreter. The executed trace finally
returns a pointer to a SIRInterpreter instance.