Optimise Yorick's internals using workspaces.
From the docstring in `xtask/src/main.rs`:

```
This is required because we need to separately compile parts of the
codebase with different configurations.

 - Interpreter code, that we expect to be able to trace, needs to be
   compiled with `-C tracer=hw`. Crucially, since hardware tracing relies
   on blocks not being reordered by LLVM, this flag disables optimisation.

 - Anything that we don't need to trace, for example the JIT internals
   themselves, can (and should) be optimised.

Note that the above definitions also extend to dependencies, and this
means that we can have two copies of any given crate but compiled with
different flags. This is good, as it means that the dependencies of
interpreter code can be traced, whereas dependencies of the JIT
internals can be optimised.

To this end, we have 2 Rust workspaces in the repo: the "internal"
(optimised) workspace, and the "external" (unoptimised) workspace. The
external workspace then talks to the internal workspace via `extern`
functions defined in the ykshim crate.
```

Note that to build/test we now use xtask. E.g.:
```
cargo xtask build
```

We also no longer need to set `RUSTFLAGS=-C tracer=hw` as xtask will do
it automatically.

Co-authored-by: bjorn3 <bjorn3@users.noreply.github.com>