Don't run step_fn after stopgapping.
This commit fixes a bug where we would execute `step_fn` once more after
running the stopgap interpreter, instead of retuning immediately
afterwards. In other words, after returning from a guard failure and
using the stopgap interpreter to get back to the control_point we would
execute the interpreter step an additional time, which can change the
outcome of a program (see the stopgapping test which without this change
would fail, as ctx.run would be incremented one more time than needed).

The fix is simple: after running the stopgap interpreter we return
immediately, in the same way we immediately return after executing a
compiled trace.