Add a 'testing' feature.
What we would like to do is mark some things (e.g. yksg's `set_interp_ctx`
function) as `#[cfg(test)]` but since we have a cross-crate testing dependency
(which is exacerbated, though not entirely caused by, the internal/external
workspace split), that doesn't work.

This commit adds a "testing" feature which is intended as yk's cross-crate
substitute for "#[cfg(test)]". It enables us to a) clearly mark some parts of
the code as only being necessary for testing b) only build those when the
"testing" feature is enabled.

This does lead to some ugliness e.g. we have to fiddle with our xtask build so
that it passes down the "testing" feature. But, overall, I think the ugliness is
worth the gain.