Try to clarify the documentation around the stopgap FramesVec.
I'm not 100% sure my updates are correct. I also think we leak the memory in
FrameInfo::mem -- we `alloc` memory but I can't see a corresponding `dealloc`.
Indeed, I think we'd need to pass size/align around for that to work but that's
rather tricky at the moment as with the x86_64 SysV ABI we only have 1 "easy"
paramater register left...