Implement trace looping.
Currently, after each trace execution we return back to the meta-tracer,
only for it to run the same trace again. A more efficient way is to make
the trace loop at the assembler level, where the only way to leave it is
via a guard failure.

The API for this very basic. The `interp_step` function is now required
to return a constant bool. If a traced path through the function returns
`true`, this means the interpreter is finished and we return; if a path
returns `false`, this means trace looping is enabled for that trace and
it can only be exited via a guard.

Typically this means interpreter authors want to do something similar to
this:

```rust
fn interp_step(io: &mut Ctx) -> bool {
    // do interpreter stuff
    if some_abort_condition {
        return true;
    }
    false // loop forever
}
```

A quick performance test has shown that this can improve the runtime of
a trace by up to 2.5x.