Implement trace stitching.
Currently, we return after each trace execution back to the meta-tracer,
only for it to run the same trace again. Instead, we can just stitch
together the compiled trace, so that it loops indefinitely until a guard
fails.

The API for this very basic. The `interp_step` function is now required
to return a bool. If a traced path through the function returns `true`,
this means the interpreter is finished and we return; if a path returns
`false`, this means trace stitching is enabled for that trace and it can
only be exited via a guard.

Typically this means interpreter authors want to do something similar to
this:

```rust
fn interp_step(io: &mut Ctx) -> bool {
    // do interpreter stuff
    if some_abort_condition {
        return true;
    }
    false // loop forever
}
```