Make debugging trace code in GDB easier.
We used to be able to set a break point for all entry points to trace
code with `b exec_trace`. Somewhere along the line this was changed and
it makes debugging harder than it needs to be (you have to step over the
call to get the trace function pointer, which is irritating).

This change fixes this and adds documentation in attempt to prevent this
regressing again.