Make the release cargo profile work.
Currently the internal workspace is always built in release mode and the
external workspace cannot be built in release mode at all (it would
build the external workspace with LLVM optimisations that hardware
tracing cannot tolerate).

This change:

 a) Enables the use of --release with `cargo xtask`. In release mode,
    the external workspace is still *not* optimised, but other release
    mode settings *are* used (e.g. unchecked operations are used, and
    debug assertions are disabled).

 b) Makes it so that the internal workspace is only built in release
    mode if `--release` is passed to xtask.

We also test both modes in CI now.

I checked that performance is unaffected when tested with --release and
that performance is worse if --release is absent (as the internal workspace
is no-longer optimised in that case).

Also checked that debug assertions can fail in the internal workspace
when using a debug build.

The only problem with this is that if software tracing is used, you
still get an unoptimised external workspace, which could in fact be
optimised. I've put a FIXME in the Cargo config to remind us the revisit
this later.