Only reserve space for registers actually in use.
The first step in our guard failure is to spill all live locals that are
in registers to the stack. This allows us to use any register for
stopgap initialisation without having to worry about overwriting live
values.

Out of simplicity (or rather laziness) we reserved space for each
register that is available to the register allocater, no matter if it
actually stored a live local or not. We thus sometimes reserved more
space than needed. To make matters worse, this code also had a bug,
which meant we were sometimes actually reserving less space than
required, since the code assumed `REG_POOL` contained all 15 registers
(it doesn't) and then used the register's integer value as an offset
into the the stack offset.

This commit kills two birds with one stone by only reserving space for
registers storing live locals, and then mapping the stack offset to the
register, to avoid this bug from happening in the future.