Use a means of probably-preventing optimisations that doesn't rely on IO.
Relying on print to force non-optimisation makes for noisy tests, and even then
it's not completely guaranteed to work (imagine if the compiler can suss out
that your code is equivalent to `if false { println!(...) }`). This commit uses
an alternative hard-to-optimise-away technique (modifying a mutable struct
that's unlikely to be easily shared between caller and callee), although, again,
there are no guarantees here.