Move away from a generic, and easy to get wrong, read_int function.
`read_int` is a type-mangling function that expects to convert any constant
integer, or an integer read from memory, to a u128. This effectively turns any
code that calls the function into dynamic typing mode, and is likely to cause
various problems (e.g. with bitwise operations).

This commit breaks out read_bool as a separate function, as it's clearly
different than "normal" integer operations. However this doesn't currently work
as, on b13 at least, it "reads" values of 210 (not 0 or 1) from a byte in one
test! This suggests that the `Terminator::Assert` code only previously worked by
accident because we effectively tested "is <0 or 1> != 210" which of course
always produced `false`.