Restructure the workspaces.
As discussed in https://github.com/softdevteam/yk/issues/311, we've
decided to renamme the 'internal' workspace to the 'untraced' workspace,
and similarly the 'external' workspace to the 'traced' workspace.

While we are here, we are also restructuring the directory hierarchy.

Whereas before we had:
```
/ (implicit external_ws)
  xtask/
  internal_ws/
    ...
  ...
```

Now we have:
```
/
  xtask/
  traced/
    ...
  untraced/
    ...
```

In other words, we have three workspaces now:
 - One (top-level) for the build system.
 - The traced workspace.
 - The untraced workspace.