Don't try to catch unwinds.
The original reason for using `catch_unwind` was to make sure we had an accurate
count of the number of active user threads. However, `catch_unwind` has
unfortunate consequences because it forces us to use `RewindSafe` types in
places where we don't want to (we can't even use `std::sync::CondVar` in `MT`,
let alone anything from the parking_lot crate!). Since we're not even sure
whether we really want unwinding or panicing, this seems too high a price to pay
-- especially because it's probably best to assume that if a thread aborts, the
whole system is probably in a broken state anyway!