Sort out usage of Ty::size and Ty::kind.
If you accessed these as attributes they were of type `usize` but if accessed as
methods `u64` which, to say the least, is confusing.

This commit removes attribute access, and makes all callers use `usize`. This
has small effects on the x86_64 code generator but, pleasingly, doesn't require
any additional try_from conversions or the like.