Move the `make_binop` macro to a less confusing place.
This was previously in the middle of unrelated code. I first moved it next to
the `impl` where it's used, and then realised that I could break out the tiny
little bit of the `impl` that's relevant to it, thus meaning that most code
readers won't need to worry about this unless they really care about binary
operations in the stopgap interpreter.