Add the ability to map PT blocks to LLVM IR blocks.
This change allows us to translate the block addresses found in a PT
trace into block identifiers in LLVM IR.

We don't yet look up the IR block in the LLVM bitcode, only collect the
information necessary to do so.

In doing so I've also created a little crate called `ykutil` which
currently only shares a routine between both `ykllvmwrap` and `yktrace`
in a sensible and dependency-cycle-free manner.

Observations:
 - LLVM blocks are not terminated by calls.
 - One PT block may map to many LLVM blocks (see doc string on
   HWTMapper::map_block() for more info).