Add a yk_stopgap_safe marker.
When adjusting an existing C interpreter, we often want to mark multiple points
in the interpreter as being safe for the stopgap interpreter to transition back
to normal "C" interpretation. This commit adds a function that we can use as
such as a marker in the future (at the moment it's a no-op). Exactly how we
teach clang/rustc/LLVM about what's special about this function is something we
can worry about in the future. Similarly, the precise name for the function is
something we can tweak in the future.

Why do we need this marker at all? Well, previously we assumed this was
synonymous with control points, but in the presence of computed gotos this isn't
really enough: interpreters rarely go through control points. So while it should
always be safe for the stopgap interpreter to transition back to normal
execution at a control point, that potentially means we execute *lots* of code
in the stopgap interpreter unnecessarily. "yk_stopgap_safe" allows an
interpreter author to add extra places in their interpreter where it's safe to
transition back to normal execution. The precise rules of "safe" are deeply
unclear, but I believe a safe, conservative approximation is "add it either
directly before a `goto` (if you're using computed gotos for your main
interpreter loop) or at the very end or beginning of your interpreter's main
loop". We might be able to relax these requirements later, but I don't think
it's something we should worry about for now.