Make IRBlock store a CString, not a String.
Currently when we hand out an IRBlock's function name to C/C++ we have
to convert the internal Rust string into a CString, before passing it to
the caller. The caller is then responsible for freeing the string. This is
unergonomic and slower than it could be.

We didn't care before as we only do this in testing, but soon we will
have some C++ code on the critical path which will need to walk an
IRTrace querying IRBlocks. Therefore we should fix this now.

Storing a CString instead of a String means that we can hand out
borrowed `char *` pointers (via CStr) without having to allocate each
time and without the caller having to free the result.