Use LLVM to compile a simple trace.
This is the first stab at our API which takes a PT trace and compiles it
using LLVM. There's still a lot missing here, so the traces we can
successfully compile are very simple.  For example, we can't yet compile
function calls like `printf` or pass values into or out of a trace.

As this may suggest, this limits us in our ability to test that the
compiled function does the right thing, so for now we only test that the
trace doesn't segfault.

Co-authored-by: Edd Barrett <vext01@gmail.com>