Improve the code that loads the raw .llvmbc section.
Currently the .llvmbc section is mmaped and loaded afresh (and never
unmapped!) for each call to IRTrace::compile().

This change makes it so that we lazily load the section data once, and
then the data (and mmap region) is reused for subsequent calls to
IRTrace::compile().

My original plan had been to do this in C++ code, but it turns out to be
quite annoying and Rust has good libraries for parsing object files.
Therefore we implement this in Rust and export the function to C++.

This also removes a lot of plumbing, which is good.