Enable passing in of values into a compiled trace.
Until now a compiled trace was a bit of a black box as we had no means
to observe what it was doing when run. This change allows us to pass
pointers into a compiled trace which we can then write to, in order to
test that the trace is executing correctly.