Inline function calls during tracing.
Since LLVM calls are not terminators, we need to be careful when
stitching together a PT trace. When we see a function call, we need to
stop copying over instructions in the current block and remember the
call instruction.

When then continue stitching as usual, which will automatically inline
the blocks of the called function, until we see a return. The next block
we see will be the same block as before inlining the call. Since we've
already copied parts of that block, we need to first skip instructions
until we reach the remembered call instruction. We then continue
stitching as normal.