Add the ability to dump a side-by-side trace to stderr.
This is a debugging aid which makes it easier to:
 - see the inlining thresholds in our JIT module.
 - see the variable mapping between the JIT and AOT module.

This feature is available only in debug builds and the dump is enabled
by setting YK_PRINT_IR_SBS=1 in the environment.

This could be improved by showing the inlined call sites (and their
arguments) in the AOT column, but one step at a time.

Co-authored-by: Lukas Diekmann <lukas.diekmann@gmail.com>