Canonicalise paths when mapping virtual addresses.
Part of the process of mapping virtual addresses back to file offsets
involves comparing object names during a search. When the binary is
invoked with a relative path, the string matching was failing for two
reasons:

1) Paths didn't match due to different (but equivalent) representations,
e.g. the absolute and relative path to the same file. Here a string
comparison doesn't suffice to decide if the paths address the same file.

2) Similarly, on Linux, some libraries go by many names due to symlinks,
so a string comparison isn't enough there either.

The solution is to canonicalise paths before comparison.

This can certainly be made faster later with caching (fs::canonicalize()
touches the file-system, and doing this repeatedly is probably slow).
I've left a PERF comment.