Add the ability to express holes in a trace.
Currently we stop mapping a trace at the first (non-prefix) unmappable
block. This is incorrect in the long run, as the trace may not be
complete. For example, it could be that there's a call to external code
for which we have no IR.

This change allows us to express holes in a trace, which we then leave
up to the trace compiler to interpret as it sees fit.