Start documenting our approach to PRs.
Although our approach is heavily influenced by projects like Rust, it's
sufficiently elaborate that newcomers find it difficult. This commit tries to
capture both the process *and* the reasons for that process. It is necessarily
incomplete and imperfect: I'm sure we'll need to iterate on this for a little
while before we feel fully happy with it. However, this is a start -- something
is better than nothing!