Don't expect fall-thru blocks to have consecutive address ranges.
In the case that a PT block maps to multiple machine blocks, it may be
tempting to check that they are at consecutive address ranges.
Unfortunately we can't do this because LLVM sometimes appends `nop`
sleds (e.g. `nop word cs:[rax + rax]; nop`) to the ends of blocks for
alignment. This padding is not reflected the LLVM block address map, so
blocks may not appear consecutive.