Use an integer instruction index to make resuming simpler.
By using an index instead of an iterator, we avoid the need to
repeatedly iterate over instructions that we want to skip: we can simply
set the index to the resume point directly.

We still need to hold on to the instruction, as it is used as a key in
the VMap. Hence we use a tuple.

Co-authored-by: Lukas Diekmann <lukas.diekmann@gmail.com>