Fix VMap element existence check.
Previously we used `VMap[Var] != nullptr` to check if an element exists
in the VMap. Unfortunately, C++ thinks its a good idea to insert a
default value (in this case `nullptr`) whenever an element is not found.
This lead to errors in LLVM which checks existence in the VMap via
`VMap.find(Var) != VMap.end()`, which now returns true after our above
check. In other words, checking if an element exists in the VMap in our
code, pushed it into existence in LLVM's world, even if it didn't
exist in the VMap before.

The solution, is to use the appropriate existence check
`VMap.find(Var) != VMap.end()` in our code. This means we also no longer
need to id-map constants to work around the `Unexpected null mapping`
error during instruction cloning.