Remap the operands of constant expressions.
Constant expressions may contain operands that require remapping, such
as global variables, but are missed because the constant expression is
simply copied over by `remapInstruction`. For example the instruction

  call foo(getelementptr i8* @str, i64 0, i64 0)

contains a constant expression `getelementptr` where one of the operands
is a global variable, which we need to copy (and remap) into JITMod.
However, while `remapInstruction` recurses into the call instruction and
updates all operands using the VMap, it does not remap the global
variable, since we haven't copied it into JITMod yet.

This commit fixes the problem, by recursively traversing all operands of
a copied instruction and making sure they are copied and remapped if
neccessary.

Normally, it would be suffucient to copy and remap the global variables
only, however, due to the way we inline functions, we also need to
recreate the constant expression: this is because inlining a function
creates a mapping `VMap[Arg] = Var`, where `Var` still points to AOTMod.
Calling `remapInstruction` on anything referencing `Arg` won't
recursively remap `Var`, since LLVM assumes anything that has an entry
in the VMap has already been fully remapped.