Mapper: Ensure we are looking up an offset from the right object.
The mapper uses `code_vaddr_to_offset()` to find the file offset of a
block. The resulting offset is then used to query the block map.

The code is incorrect because `code_vaddr_to_offset()` also succeeds on
externally linked code such as shared objects, yet the blockmap only
describes blocks in the main binary, thus sometimes we look up an offset
for the wrong file.

Most of the time this is harmless, as block offsets in different objects
rarely collide, and thus looking up a foreign offset in the blockmap
gives no result anyway. However, when experimenting with tracing the
Ruby interpreter, we did observe a collision, leading to
`find_code_sym()` to fail and in turn the JIT to crash.

The correct thing to do (until we can load a blockmap and IR for each
object) is to check that the offset in question comes from the main
binary and failing that report that there's external code in the trace
(return an empty vector).

Unfortunately this is very hard to devise a test for (we'd need objects
with colliding block addresses).