Fix indirectbr test for AOT opt levels > -O0.
The indirect branch test uses computed gotos, as often seen in threaded
dispatch in interpreters.

The C code `goto *dispatch[argc];` (at higher AOT optimisation levels)
is realised as IR like:

```
@__const.main.dispatch = private unnamed_addr constant [3 x i8*] [i8* blockaddress(@main, %12), i8* blockaddress(@main, %13), i8* blockaddress(@main, %14)], align 16

define dso_local i32 @main(i32 %0, i8** nocapture readnone %1) #0 !dbg !16 {
  ...
  %10 = getelementptr inbounds [3 x i8*], [3 x i8*]* @__const.main.dispatch, i64 0, i64 %9, !dbg !52
  %11 = load i8*, i8** %10, align 8, !dbg !52, !tbaa !53
  indirectbr i8* %11, [label %12, label %13, label %14]
  ...
}
```

Thus, there are several intermediate IR instructions and a global array
involved in setting up the `indirectbr` instruction.

When we create the trace code, It isn't necessary to copy the indirect
branch into the `JITMod` as the successor block is known from the trace.
However, naively not copying the branch would lead to dangling
references in the IR because the `address` operand typically
(indirectly) references AOT block addresses not present in the `JITMod`.
Therefore we also remove the IR instruction which defines the `address`
operand and anything which also becomes dead as a result (recursively).