Deal with indirect calls.
This implements ways to deal with indirect function calls, i.e. calls
where the callee isn't statically known, e.g.

%1 = @foo
call %1

Since at the time we trace the call instruction we can't tell what %1
is, we need to look ahead in the trace which function the tracer saw
next.

There's two scenarios we have to deal with: if we have IR for the passed
function, we can inline it as usual (note that in the future we don't
want to do this unless the function was promoted); if we don't have IR,
we can't inline, so we leave the call as is (we do however need to make
sure that the called function has been redeclared in the JITModule).