Fix incorrect deletion of potentially "used later" values.
The change in 8a9928354d deletes known dead values recursively if their
sole user is also going to be deleted.

This didn't take into account that when we do the deletions, we are in
the middle of building the JIT module, and that further users may appear
later. Thus, we may erroneously delete an operand that is actually
required.

This change amends the indirect call test to trigger this bug and fixes
it by caching the dead values until finalisation of the JITMod. Only
then do we consider deleting dead values.