Encapsulate the function address fields in a type.
This hides the low-level details of these arrays and exposes them in a
more pleasing way.

We also use a hashmap internally to avoid linearly scanning.

In theory this might improve performance a bit, as it avoids repeated
string comparisons. In reality we (currently) query the map so rarely,
so it's unlikely to make a big difference, but at least it makes the
code a bit more consise.