Hide the thread tracer from the consumer.
This change hides the underlying `ThreadTracer` inside a thread local.

Before we would return the thread tracer to the code that is being
traced when we starting tracing. This meant that code that dealt with
shuffling the `ThreadTracer` around ended up in the recorded trace, yet
we really don't care for such code. Worse, its presence complicates
things like computing the trace inputs (an upcoming unit of work).

Rather than trying to apply ad-hoc hacks to strip this unwanted noise
from the trace, we simply hide the `ThreadTracer` inside the yk library,
never exposing it to consumers of the library. Then it can never appear
in a trace.