Implement support for the new control point design.
The new control point design replaces a call to a dummy control point
with a dynamically generated control point.

This new control point accepts and returns a struct, whose members are
derived from the live state of the interpreter just before the control
point.

Generation of the control point is implemented in ykllvm. This change
implements the remainder of the logic (namely JIT runtime support).

Currently only a few tests run. In order to get those working we had to
fix a bug relating to PHI nodes.

Fixes #394

Co-authored-by: Edd Barrett <vext01@gmail.com>