Use intrinsic annotations to correctly assemble traces.
Since LLVM may decide to entirely remove intrinsic calls, its impossible
to reliably map a PT trace back to LLVM IR. In other words, if an
intrinsic stays a call we see the same block twice, if it was inlined we
don't.

With our recent change that annotates intrinsic calls with inlining
decisions, we can make the right decision during trace construction:
e.g. if the intrinsic was inlined we don't need to expect a dpulicated
block and can just move on.