Implement aborting guards for simple branch instructions.
When the trace compiler encounters an LLVM `br` instruction in a trace,
it now emits a guard with the appropriate check.

For now a failing guard simply branches to a block which aborts the
program (eventually a guard failure should invoke a stop-gap
interpreter, but we are not ready for that yet).

This of course now means that the `JITFunc` is no longer one linear
block, but that may change in the future, depending on how we decide to
transfer to the stop-gap interpreter.

Without implementing trace stitching, there is currently no way for a
test to exit with success, so for now all of the (unskipped) tests now
expect to crash.

I've strengthened the existing tests, including checking for emission of
the guard blocks.

I also had to change the lang test name matchers so that we could match
a basicblock *value* (prefixed with `%`) to its *label* (without the
prefix).

Soon we should implement guards for LLVM `switch` and `indirectbr`
instructions too, but those should come in future PRs.