Move state transition code into Rust code.
This change moves some of the (compile-time generated) control point
logic into Rust code. Functionally this change is a no-op.

The plan is to keep as little code in compile-time generated LLVM IR, as
it's quite laborious to write and experiment with. Also we are about to
grapple with concurrency, and doing that in compile-time generated IR
would be a nightmare.

Some of the design introduced in this change is pretty questionable,
e.g. storing pointers in Rust globals using `usize`. I plan for this to
be eliminated when we properly hook in `transition_location()`, but I
feel it would be foolish to not use this interim PR as a stepping stone.